pyproject.toml
[project]
name = "sleephq-client"
version = "0.1.0"
description = "Python SDK for the SleepHQ API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.23.0,<0.29.0",
    "attrs>=22.2.0",
    "python-dateutil>=2.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sleephq"]

[dependency-groups]
dev = ["pytest>=7", "mypy>=1.0", "types-python-dateutil>=2.8", "openapi-python-client>=0.21"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
